+++ app/Facades/Experiment.php
@@ -2,6 +2,7 @@
 
 namespace App\Facades;
 
+use App\Services\Experiment as ExperimentService;
 use Illuminate\Support\Facades\Facade;
 
 /*
@@ -10,13 +11,8 @@ use Illuminate\Support\Facades\Facade;
  */
 class Experiment extends Facade
 {
-    public static function is($name)
+    protected static function getFacadeAccessor()
     {
-        return static::$app['request']->session()->get('experiment') === $name;
-    }
-
-    public static function set($name)
-    {
-        return static::$app['request']->session()->put('experiment', $name);
+        return ExperimentService::class;
     }
 }