+++ resources/views/harvests/index.blade.php
@@ -61,15 +61,15 @@ Spice Harvester |
                             <td>{!! $h->cron_status !!}</td>
 			                <td class="text-right">
                                 {!! Form::open(array('method' => 'DELETE', 'route' => array('harvests.destroy', $h->id), 'class' => 'visible-xs-inline form-inline')) !!}
-                                {!! link_to_action('SpiceHarvesterController@show', 'Detail', array($h->id), array('class' => 'btn btn-primary btn-detail btn-xs btn-outline', )) !!}
-                                {!! link_to_action('SpiceHarvesterController@edit', 'Upraviť', array($h->id), array('class' => 'btn btn-primary btn-xs btn-outline')) !!}
+                                {!! link_to_action('App\Http\Controllers\SpiceHarvesterController@show', 'Detail', array($h->id), array('class' => 'btn btn-primary btn-detail btn-xs btn-outline', )) !!}
+                                {!! link_to_action('App\Http\Controllers\SpiceHarvesterController@edit', 'Upraviť', array($h->id), array('class' => 'btn btn-primary btn-xs btn-outline')) !!}
                                     {!! Form::submit('Zmazať', array('class' => 'btn btn-danger btn-xs btn-outline')) !!}
                                 {!! Form::close() !!}
                                 <br>
-                                {!! link_to_action('SpiceHarvesterController@orphaned', 'Zobraziť odobrané zo setu', array($h->id), array('class' => 'btn btn-danger ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
-                                {!! link_to_action('SpiceHarvesterController@launch', 'Obnoviť všetky', array($h->id, 'reindex'=>true), array('class' => 'btn btn-default ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
-                                {!! link_to_action('SpiceHarvesterController@harvestFailed', 'Obnoviť chybné (' . $h->records()->failed()->count() . ')', array($h->id), array('class' => 'btn btn-default ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
-                                {!! link_to_action('SpiceHarvesterController@launch', 'Spustiť', array($h->id), array('class' => 'btn btn-primary ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
+                                {!! link_to_action('App\Http\Controllers\SpiceHarvesterController@orphaned', 'Zobraziť odobrané zo setu', array($h->id), array('class' => 'btn btn-danger ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
+                                {!! link_to_action('App\Http\Controllers\SpiceHarvesterController@launch', 'Obnoviť všetky', array($h->id, 'reindex'=>true), array('class' => 'btn btn-default ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
+                                {!! link_to_action('App\Http\Controllers\SpiceHarvesterController@harvestFailed', 'Obnoviť chybné (' . $h->records()->failed()->count() . ')', array($h->id), array('class' => 'btn btn-default ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
+                                {!! link_to_action('App\Http\Controllers\SpiceHarvesterController@launch', 'Spustiť', array($h->id), array('class' => 'btn btn-primary ladda-button btn-xs', 'data-style'=>'expand-right')) !!}
                             </td>
 			            </tr>
 						@endforeach